refactor: Moved package-info back to utils
