+++ src/internal.ts
@@ -19,7 +19,7 @@ export {
   isESMExport,
   type ExportCondition,
   type ExportEntry,
-} from './lib/package-resolution/package-info.js';
+} from './lib/utils/package/package-info.js';
 export { isInSkipList, prepareSkipList } from './lib/config/default-skip-list.js';
 
 export { NfFileWatcher, NfFileWatcherOptions } from './lib/domain/utils/file-watcher.contract.js';
