+++ src/lib/utils/package/esm-detection.spec.ts
